通过拦截器 和 JWT 实现 管理员是否有操作删除权限。

+++ blog_user/src/main/java/com/dongl/user/service/UserService.java
@@ -6,11 +6,10 @@ import com.dongl.user.config.BCryptUtil;
 import com.dongl.user.dao.UserDao;
 import com.dongl.user.entity.User;
 import com.dongl.user.rabbitmq.SendMsgService;
-import com.dongl.user.rabbitmq.SendMsgServiceImpl;
 import com.dongl.utils.IdWorker;
+import com.dongl.utils.JwtUtil;
+import com.dongl.utils.StrUtils;
 import org.apache.commons.lang3.RandomStringUtils;
-import org.apache.commons.lang3.StringUtils;
-import org.springframework.amqp.rabbit.core.RabbitTemplate;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -22,6 +21,7 @@ import javax.persistence.criteria.CriteriaBuilder;
 import javax.persistence.criteria.CriteriaQuery;
 import javax.persistence.criteria.Predicate;
 import javax.persistence.criteria.Root;
+import javax.servlet.http.HttpServletRequest;
 import java.util.*;
 
 /**
@@ -47,6 +47,9 @@ public class UserService {
 
     @Autowired
     private IdWorker idWorker;
+
+    @Autowired
+    private HttpServletRequest request;
 
     /**
      * 查询全部列表
@@ -126,11 +129,14 @@ public class UserService {
     }
 
     /**
-     * 删除
-     *
+     * 删除用户时，需要登录切有相应的权限
      * @param id
      */
     public void deleteById(String id) {
+        Object claims_admin = request.getAttribute("claims_admin");
+        if (StrUtils.isEmptyObj(claims_admin)){
+            throw new RuntimeException("权限不足");
+        }
         userDao.deleteById(id);
     }
